+++ front/src/component/publish/PublishForm.jsx
@@ -1,25 +1,16 @@
 import React, { useEffect, useState, useRef } from 'react';
 import styled from 'styled-components';
-import { useNavigate, useLocation } from 'react-router-dom';
-import { useSelector, useDispatch } from 'react-redux';
-import categoryData from './categoryData';
-import { DefaultButton, TransparentButton } from '../common/button/ButtonStyle';
+import { useLocation } from 'react-router-dom';
 import PublishPhoto from './PublishPhoto';
-import ConfirmModal from '../common/modal/ConfirmModal';
-import YesNoModal from '../common/modal/YesNoModal';
-import publishApi from '../../api/publishApi';
-import { loginModalActions } from '../../redux/loginModalSlice';
-import postEditApi from '../../api/postEditApi';
+import PublishLocCat from './PublishLocCat';
+import PublishTags from './PublishTags';
+import PublishModalButton from './PublishModalButton';
 
 function PublishForm() {
-  const navigate = useNavigate();
-  const dispatch = useDispatch();
   const loc = useLocation().state;
   const isPublishPage = loc === null; // 작성/수정페이지 구분
   const ref = useRef(); // 자식 컴포넌트 ref 설정
-
-  const login = useSelector(state => state.login.isLogin);
-
+  const [boardId, setBoardId] = useState(); // 수정 post보낼 board id
   const [formData, setFormData] = useState({
     title: '',
     content: '',
@@ -31,13 +22,7 @@ function PublishForm() {
 
   const [photoUrl, setPhotoUrl] = useState(); // S3에서 가져온 URL정보
   const [images, setImages] = useState([]); // URL정보 보관
-  const [categorySelected, setCategorySelected] = useState('');
   const [tags, setTags] = useState([]);
-  const [newTag, setNewTag] = useState();
-  const [boardId, setBoardId] = useState(); // 수정 post보낼 board id
-
-  const [confirmModalOpened, setConfirmModalOpened] = useState(false);
-  const [yesNoModalOpened, setYesNoModalOpened] = useState(false);
 
   // 유효성검사
   const [titleValid, setTitleValid] = useState(false);
@@ -50,7 +35,6 @@ function PublishForm() {
     const InputName = event.target.name;
     const InputValue = event.target.value;
     const InputLength = InputValue.length;
-
     if (InputName === 'title') {
       if (InputLength < 5) {
         setTitleMessage('5글자 이상 입력하세요');
@@ -67,117 +51,8 @@ function PublishForm() {
         setContentMessage('5글자 이상 입력하세요');
         setContentValid(false);
       } else setContentValid(true);
-
     setFormData({ ...formData, [event.target.name]: event.target.value });
   };
-
-  const mandatoryInfo =
-    images.length !== 0 &&
-    titleValid &&
-    contentValid &&
-    formData.category !== undefined;
-
-  // 테마 선택 함수
-  const onClick = index => {
-    setCategorySelected(index);
-    const { category } = categoryData[index];
-    setFormData({ ...formData, category });
-  };
-
-  // 해쉬태그 추가
-  const handleTagAdd = async event => {
-    const { value } = event.target;
-    const filtered = tags.filter(el => el === value);
-
-    if (event.key === 'Enter' || event.key === ',') {
-      if (value !== '' && filtered.indexOf(value) === -1 && value.length <= 6) {
-        if (tags.length <= 4) {
-          setTags([...tags, value]);
-          setFormData({ ...formData, tags: [...tags, value] });
-          setTimeout(() => {
-            setNewTag('');
-          }, 0);
-        }
-      }
-    }
-  };
-
-  // 해쉬태그 삭제
-  const handleTagRemove = indexRemove => {
-    setTags([...tags.filter((_, index) => index !== indexRemove)]);
-  };
-
-  // 게시글 등록 요청
-  const publishRequest = async event => {
-    event.preventDefault();
-    if (mandatoryInfo)
-      await publishApi(formData)
-        .then(res => {
-          if (res.status === 201) {
-            console.log(res);
-            console.log(res.data.id);
-            setConfirmModalOpened(true);
-            setTimeout(() => {
-              navigate(`/postDetail/${res.data.id}`);
-            }, 2000);
-          }
-        })
-        .catch(error => console.log(error.response.data.message));
-    else console.log('not valid'); // 임시
-  };
-
-  // 게시글 수정 요청
-  const editRequest = async event => {
-    event.preventDefault();
-    if (
-      images.length !== 0 &&
-      formData.title.length >= 5 &&
-      formData.content.length >= 5
-    )
-      await postEditApi(boardId, formData)
-        .then(res => {
-          if (res.status === 200) {
-            setConfirmModalOpened(true);
-            setTimeout(() => {
-              navigate(`/postDetail/${res.data.id}`);
-            }, 1000);
-          }
-        })
-        .catch(error => console.log(error.response.data.message));
-  };
-
-  // 모달 닫는 함수
-  const confirmModalCloser = () => {
-    setConfirmModalOpened(false);
-  };
-
-  // 취소 버튼 모달 연결
-  const yesNoModalOpener = event => {
-    event.preventDefault();
-    setYesNoModalOpened(true);
-  };
-
-  const yesNoModalActioner = () => {
-    navigate(-1);
-  };
-
-  const yesNoModalCloser = () => {
-    setYesNoModalOpened(false);
-  };
-
-  const loginModalOpener = () => {
-    dispatch(loginModalActions.openLoginModal());
-  };
-
-  // (비로그인) url입력 접근시 로그인창으로 redirect
-  useEffect(() => {
-    if (!login) {
-      alert('로그인 해주세요');
-      navigate('/');
-      loginModalOpener();
-    }
-    return () => {};
-  }, []);
 
   // 사진정보 formData에 담아주기
   useEffect(() => {
@@ -220,6 +95,14 @@ function PublishForm() {
       ref.current.preview([...data.photos]); // 장착시 미리보기 실행 코드
     }
   }, [loc]);
+
+  console.log(formData);
+  const mandatory =
+    images.length !== 0 &&
+    formData.title.length >= 5 &&
+    formData.title.length <= 40 &&
+    formData.content.length >= 5 &&
+    formData.category !== undefined;
 
   return (
     <Container>
@@ -271,108 +154,25 @@ function PublishForm() {
           {contentValid ? null : contentMessage}
         </span>
       </ContentContainer>
-      <LocationCategoryRow>
-        <LocationContainer>
-          <label className="location__label" htmlFor="location">
-            위치
-          </label>
-          <input
-            className="location__input"
-            type="text"
-            id="location"
-            name="location"
-            defaultValue={isPublishPage ? null : loc.post.location}
-            onChange={event => onChange(event)}
-            maxLength="40"
-            placeholder="위치를 남겨주세요"
-          />
-        </LocationContainer>
-        <CategoryContainer>
-          <label htmlFor="category">테마</label>
-          <div id="categories">
-            {categoryData.map((item, index) => {
-              return (
-                <Category
-                  theme={item.theme}
-                  onClick={() => {
-                    onClick(index);
-                  }}
-                  className={
-                    categorySelected === index ? 'categorySelected' : null
-                  }
-                  key={item.id}
-                >
-                  {item.text}
-                </Category>
-              );
-            })}
-          </div>
-        </CategoryContainer>
-      </LocationCategoryRow>
-      <TagContainer>
-        <label htmlFor="tag">태그</label>
-        <ul className="tag__wrapper">
-          {tags.map((tag, index) => (
-            <li className="tag" key={tag}>
-              <span className="tag__name">#{tag}</span>
-              <button onClick={() => handleTagRemove(index)}>❌</button>
-            </li>
-          ))}
-          <input
-            className="tag__input--create"
-            id="tag"
-            type="text"
-            value={newTag || ''}
-            maxLength="6"
-            onChange={event => setNewTag(event.target.value)}
-            onKeyDown={event => handleTagAdd(event)}
-            placeholder="# 태그를 입력하세요"
-            disabled={tags.length === 5}
-          />
-        </ul>
-      </TagContainer>
-      <ButtonContainer>
-        <PublishButton
-          width="8vw"
-          height="4vh"
-          fontSize="var(--font-15)"
-          fontWeight="var(--font-bold)"
-          onClick={isPublishPage ? publishRequest : editRequest}
-        >
-          <span>{isPublishPage ? '등록' : '수정'}</span>
-        </PublishButton>
-        <CancelButton
-          width="8vw"
-          fontSize="var(—font-15)"
-          fontWeight="var(—font-bold)"
-          onClick={yesNoModalOpener}
-        >
-          <span>취소</span>
-        </CancelButton>
-      </ButtonContainer>
-      <>
-        {confirmModalOpened ? (
-          <ConfirmModal
-            modalMessage={
-              isPublishPage
-                ? '게시글 등록이 완료되었습니다.'
-                : '게시글 수정이 완료되었습니다.'
-            }
-            modalCloser={confirmModalCloser}
-          />
-        ) : null}
-        {yesNoModalOpened ? (
-          <YesNoModal
-            modalMessage={
-              isPublishPage
-                ? '게시글 작성을 취소할까요?'
-                : '게시글 수정을 취소할까요?'
-            }
-            modalActioner={yesNoModalActioner}
-            modalCloser={yesNoModalCloser}
-          />
-        ) : null}
-      </>
+      <PublishLocCat
+        isPublishPage={isPublishPage}
+        loc={loc}
+        onChange={onChange}
+        formData={formData}
+        setFormData={setFormData}
+      />
+      <PublishTags
+        formData={formData}
+        setFormData={setFormData}
+        tags={tags}
+        setTags={setTags}
+      />
+      <PublishModalButton
+        boardId={boardId}
+        mandatory={mandatory}
+        formData={formData}
+        isPublishPage={isPublishPage}
+      />
     </Container>
   );
 }
@@ -454,136 +254,5 @@ const ContentContainer = styled.section`
     @media screen and (max-width: 549px) {
       height: 10rem;
     }
-  }
-`;
-
-// 위치 & 테마 열
-const LocationCategoryRow = styled.section`
-  display: flex;
-  justify-content: space-between;
-  width: 100%;
-  gap: 1.5rem;
-  @media screen and (max-width: 549px) {
-    flex-direction: column;
-  }
-`;
-
-// 위치 - 입력(선택)
-const LocationContainer = styled.div`
-  width: 100%;
-`;
-
-// 테마 - 선택(필수:택1)
-const CategoryContainer = styled.div`
-  width: 100%;
-
-  #categories {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-around;
-    border: 1px solid var(--holder-base-color);
-    border-radius: var(--radius-10);
-    padding: 0.5rem;
-  }
-  @media screen and (max-width: 549px) {
-    width: 70%;
-  }
-`;
-
-const Category = styled.button`
-  background-color: ${props => props.theme.background};
-  color: ${props => props.theme.color};
-  text-align: center;
-  width: 30%;
-  border: none;
-  border-radius: var(--radius-10);
-  font-weight: var(--font-semi-bold);
-  font-size: 13px;
-  opacity: 0.7;
-  padding: 0.5rem;
-  cursor: pointer;
-  &:hover,
-  &.categorySelected {
-    opacity: 1;
-    font-weight: var(--font-bold);
-    transition: 0.2s all ease-in-out;
-    box-shadow: var(--bx-sh-four);
-  }
-  @media screen and (max-width: 549px) {
-    font-size: 10px;
-  }
-`;
-
-// 태그 - 입력, 삭제
-const TagContainer = styled.section`
-  width: auto;
-  .tag__wrapper {
-    display: flex;
-    flex-direction: row;
-    max-width: 100%;
-    height: auto;
-    overflow: hidden;
-    border-radius: var(--radius-10);
-    border: 1px solid var(--holder-base-color);
-    flex-wrap: wrap;
-    .tag {
-      display: flex;
-      align-items: center;
-      white-space: nowrap;
-      margin: 0 0.5rem 0 1rem;
-      border-radius: var(--radius-10);
-      color: var(--font-base-black);
-      font-size: 1.5rem;
-      > button {
-        text-align: center;
-        font-size: 1rem;
-        opacity: 0.8;
-        background-color: transparent;
-        border: none;
-        margin-left: 3px;
-        color: var(--font-base-black);
-        cursor: pointer;
-        &:hover {
-          scale: calc(1.3);
-          transition: 0.1s all linear;
-        }
-      }
-    }
-    .tag__input--create {
-      border: none;
-      flex: 1;
-      margin-top: 0;
-    }
-  }
-`;
-
-// 버튼 - 등록, 취소
-const ButtonContainer = styled.form`
-  display: flex;
-  justify-content: start;
-`;
-
-// 등록 버튼 스타일링
-const PublishButton = styled(DefaultButton)`
-  @media screen and (max-width: 549px) {
-    width: 25vw;
-    height: 4vh;
-  }
-  &:hover {
-    background: var(—-button-theme-hv);
-    color: var(—-base-white-color);
-    transition: 0.1s ease-in-out;
-  }
-`;
-
-// 취소 버튼 스타일링
-const CancelButton = styled(TransparentButton)`
-  &:hover {
-    color: red;
-    transition: 0.1s ease-in-out;
-  }
-  @media screen and (max-width: 549px) {
-    width: 28vw;
-    height: 4vh;
   }
 `;
